cli.py
#!/usr/bin/env python3
"""
HERALD v1.0 Command Line Interface
CPU-optimized AI architecture for local deployment

Usage:
    python cli.py serve [--host HOST] [--port PORT]
    python cli.py load-model MODEL_PATH
    python cli.py generate TEXT [--max-tokens N]
    python cli.py benchmark
"""

import argparse
import sys
import yaml
from pathlib import Path
from typing import Optional

# Import HERALD components (will be implemented later)
# from core.engine import NeuroEngine
# from core.tokenizer import ASCTokenizer
# from api.server import HeraldServer


def load_config(config_path: str = "config/runtime_config.yaml") -> dict:
    """Load configuration from YAML file."""
    try:
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    except FileNotFoundError:
        print(f"Warning: Configuration file {config_path} not found. Using defaults.")
        return {}


def serve_command(args):
    """Start the HERALD API server."""
    config = load_config()
    
    host = args.host or config.get('server', {}).get('host', '0.0.0.0')
    port = args.port or config.get('server', {}).get('port', 8000)
    
    print(f"Starting HERALD server on {host}:{port}")
    print("API documentation available at http://localhost:8000/docs")
    
    # TODO: Implement server startup
    # server = HeraldServer(config)
    # server.start(host=host, port=port)
    
    print("Server startup not yet implemented.")


def load_model_command(args):
    """Load a HERALD model from file."""
    model_path = Path(args.model_path)
    
    if not model_path.exists():
        print(f"Error: Model file {model_path} not found.")
        sys.exit(1)
    
    print(f"Loading model from {model_path}")
    
    # TODO: Implement model loading
    # engine = NeuroEngine()
    # engine.load_model(model_path)
    
    print("Model loading not yet implemented.")


def generate_command(args):
    """Generate text using HERALD."""
    text = args.text
    max_tokens = args.max_tokens or 100
    
    print(f"Generating {max_tokens} tokens for: {text[:50]}...")
    
    # TODO: Implement text generation
    # engine = NeuroEngine()
    # result = engine.generate(text, max_tokens=max_tokens)
    
    print("Text generation not yet implemented.")


def benchmark_command(args):
    """Run performance benchmarks."""
    print("Running HERALD performance benchmarks...")
    
    # TODO: Implement benchmarking
    # benchmarks = BenchmarkSuite()
    # results = benchmarks.run_all()
    
    print("Benchmarking not yet implemented.")


def main():
    """Main CLI entry point."""
    parser = argparse.ArgumentParser(
        description="HERALD v1.0 - CPU-optimized AI architecture",
        formatter_class=argparse.RawDescriptionHelpFormatter,
        epilog="""
Examples:
  python cli.py serve --port 8080
  python cli.py load-model ./models/herald-v1.0.herald
  python cli.py generate "Hello, world!" --max-tokens 50
  python cli.py benchmark
        """
    )
    
    subparsers = parser.add_subparsers(dest='command', help='Available commands')
    
    # Serve command
    serve_parser = subparsers.add_parser('serve', help='Start the API server')
    serve_parser.add_argument('--host', default='0.0.0.0', help='Host to bind to')
    serve_parser.add_argument('--port', type=int, default=8000, help='Port to bind to')
    serve_parser.set_defaults(func=serve_command)
    
    # Load model command
    load_parser = subparsers.add_parser('load-model', help='Load a HERALD model')
    load_parser.add_argument('model_path', help='Path to .herald model file')
    load_parser.set_defaults(func=load_model_command)
    
    # Generate command
    generate_parser = subparsers.add_parser('generate', help='Generate text')
    generate_parser.add_argument('text', help='Input text to continue')
    generate_parser.add_argument('--max-tokens', type=int, default=100, 
                               help='Maximum tokens to generate')
    generate_parser.set_defaults(func=generate_command)
    
    # Benchmark command
    benchmark_parser = subparsers.add_parser('benchmark', help='Run benchmarks')
    benchmark_parser.set_defaults(func=benchmark_command)
    
    args = parser.parse_args()
    
    if not args.command:
        parser.print_help()
        sys.exit(1)
    
    try:
        args.func(args)
    except KeyboardInterrupt:
        print("\nOperation cancelled by user.")
        sys.exit(0)
    except Exception as e:
        print(f"Error: {e}")
        sys.exit(1)


if __name__ == '__main__':
    main() 